refactor: update README links and add network dashboard script for video generation testing
